requirements.txt
asgiref==3.7.2
Django==5.0.3
psycopg==3.1.18
psycopg-binary==3.1.18
sqlparse==0.4.4
django-allauth==0.61.1
pytest==8.1.1
pytest-django==4.8.0
factory-boy==3.3.0
black==24.2.0
